shortcut-fe : changes to skip mtu check at SFE
In cases where WAN MTU is larger than LAN MTU,
APPS does the fragmentation and fragmented packets
are given to Q6. This results in a overhead and
increase in MIPS. As Q6 already has the logic to
do fragmentation, introduce a new proc entry to
skip the mtu processing at APPS. This results in
MIPS saving at APPS and in overall fragmentation
numbers will increase.

Change-Id: I21a9ced704c6d338f22ed2bedb6895facfebffda