shortcut-fe: Change to remove unnecessary verification of session_idx
verification of session_idx is unnecessary as it is already done
at the driver level of l2tp before conntrack is made. This change
is needed to remove the check for session_idx

Also, this change addresses a bug, which was making the packets
to skip SFE and route through normal data path.

Change-Id: Iab3fef3fcdf7460a69453dfb62f7fb1245992044
Signed-off-by: Aman Gupta <amangupt@codeaurora.org>