Add Timer Deferrable flag in timer setup.
Non-deferrable timer leading to APPS wake for every 10ms during s2idle suspend so added TIMER_DEFERRABLE flag as part of timer setup.

TIMER_DEFERRABLE: A deferrable timer will work normally when the
system is busy, but will not cause a CPU to come out of idle just
to service it; instead, the timer will be serviced when the CPU
eventually wakes up with a subsequent non-deferrable timer.

Change-Id: Ib4ced11965067ead5bf17c11e0ec3efa728568d2
CRs-Fixed: 4019951